Setup user with planner shopping list Create planner shopping list if it doesn't exist in planner

+++ app/helpers/planner_shopping_list_helper.rb
@@ -125,16 +125,17 @@ module PlannerShoppingListHelper
 	end
 
 	def update_planner_shopping_list_portions
-		if current_user.planner_recipes.length == 0
+		planner_shopping_list = PlannerShoppingList.find_or_create_by(user_id: current_user.id)
+		if current_user.planner_recipes && current_user.planner_recipes.length == 0
 			current_user.combi_planner_shopping_list_portions.destroy_all
-			current_user.planner_shopping_list.update_attributes(
+			planner_shopping_list.update_attributes(
 				ready: true
 			)
 			return
 		end
 		current_user.combi_planner_shopping_list_portions.select{|cp|cp.planner_shopping_list_portions.length == 0 || cp.planner_shopping_list_portions.length == 1 }.map{|cp| cp.destroy}
 
-		planner_shopping_list = PlannerShoppingList.find_or_create_by(user_id: current_user.id)
+
 		planner_shopping_list_portions = planner_shopping_list.planner_shopping_list_portions.select{|p|p.planner_recipe.date >= Date.current}.sort_by{|p|p.planner_recipe.date}
 		ingredients_from_planner_shopping_list = planner_shopping_list_portions.map(&:ingredient_id)
